requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
lxml==3.7.3
olefile==0.44
Pillow==4.1.1
PyInstaller==3.2.1
pytesseract==0.1.7
selenium==3.4.3
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.2.0
